+++ fortpy/__init__.py
@@ -1,5 +1,5 @@
 __all__ = ["code","elements", "parsers"]
-__version__ = "1.7.1"
+__version__ = "1.7.2"
 from . import elements
 import fortpy.parsers
 from . import code