+++ tests/isense/original/monte_orderparam.f90
@@ -309,7 +309,7 @@ contains
        close(this%fileunit)
     end if
     call pysave(
-    call pysave('dog.cat', 8,
+    call pysave('dog.cat', 8
     call py
     call pysave(fir
     call rand